Basic fluid simulation in, Split module
