+++ src/chunk.rs
@@ -1,6 +1,7 @@
 use crate::{
-    particle::{Fluid, Particle, ParticleType},
+    particle::{Particle, ParticleType},
     render::chunk_material::INDICE_BUFFER_SIZE,
+    simulation::fluid::simulate_fluid,
 };
 use bevy::{prelude::*, utils::HashMap};
 
@@ -168,152 +169,5 @@ impl Chunk {
     /// Checks if the given local position is within chunk bounds
     pub fn is_in_bounds(&self, local_pos: UVec2) -> bool {
         local_pos.x < CHUNK_SIZE && local_pos.y < CHUNK_SIZE
-    }
-}
-
-/// Checks if the given coordinates are within the bounds of a chunk
-fn is_position_valid(x: i32, y: i32) -> bool {
-    x >= 0 && x < CHUNK_SIZE as i32 && y >= 0 && y < CHUNK_SIZE as i32
-}
-
-/// Checks if a position is available (empty) in both original and new cells
-fn is_cell_available(
-    original_cells: &[[Option<Particle>; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
-    new_cells: &[[Option<Particle>; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
-    x: usize,
-    y: usize,
-) -> bool {
-    original_cells[x][y].is_none() && new_cells[x][y].is_none()
-}
-
-/// Checks if a position is valid and available in both original and new cells
-fn is_position_valid_and_available(
-    original_cells: &[[Option<Particle>; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
-    new_cells: &[[Option<Particle>; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
-    x: i32,
-    y: i32,
-) -> bool {
-    // First check bounds to avoid invalid conversions to usize
-    if !is_position_valid(x, y) {
-        return false;
-    }
-
-    // Convert to usize only after bounds check
-    let x_usize = x as usize;
-    let y_usize = y as usize;
-
-    // Check if cell is available
-    original_cells[x_usize][y_usize].is_none() && new_cells[x_usize][y_usize].is_none()
-}
-
-/// Calculates the new position for a fluid particle, reading from original_cells and writing to new_cells
-fn simulate_fluid(
-    original_cells: &[[Option<Particle>; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
-    new_cells: &mut [[Option<Particle>; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
-    fluid: Fluid,
-    x: u32,
-    y: u32,
-) {
-    let buoyancy = fluid.get_buoyancy();
-    let viscosity = fluid.get_viscosity();
-
-    // Skip if buoyancy is 0 (no movement)
-    if buoyancy == 0 {
-        // Keep the fluid in place
-        new_cells[x as usize][y as usize] = Some(Particle::Fluid(fluid));
-        return;
-    }
-
-    // Determine the vertical direction and check boundaries
-    let new_y_i32 = y as i32 + buoyancy * viscosity;
-    // Use max to ensure new_y is at least 0
-    let new_y = (new_y_i32.max(0)) as u32;
-
-    // Check if we're at a vertical boundary
-    if !is_position_valid(x as i32, new_y as i32) {
-        // Keep the fluid in place at the boundary
-        new_cells[x as usize][y as usize] = Some(Particle::Fluid(fluid));
-        return;
-    }
-
-    let new_y = new_y as usize;
-
-    // Try to move in one of three directions based on priority
-    // Note: We check the original cells for obstacles, but write to new_cells
-    if is_cell_available(original_cells, new_cells, x as usize, new_y) {
-        // Move vertically
-        new_cells[x as usize][new_y] = Some(Particle::Fluid(fluid));
-    } else if x > 0
-        && is_position_valid_and_available(original_cells, new_cells, (x - 1) as i32, new_y as i32)
-    {
-        // Move diagonally to the left
-        new_cells[(x - 1) as usize][new_y] = Some(Particle::Fluid(fluid));
-    } else if is_position_valid_and_available(
-        original_cells,
-        new_cells,
-        (x + 1) as i32,
-        new_y as i32,
-    ) {
-        // Move diagonally to the right
-        new_cells[(x + 1) as usize][new_y] = Some(Particle::Fluid(fluid));
-    } else {
-        // If fluid can't move in any of these directions, it stays in place
-        new_cells[x as usize][y as usize] = Some(Particle::Fluid(fluid));
-    }
-}
-
-/// Calculates the new position for a sand particle, reading from original_cells and writing to new_cells
-#[expect(dead_code)]
-fn simulate_sand(
-    original_cells: &[[Option<Particle>; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
-    new_cells: &mut [[Option<Particle>; CHUNK_SIZE as usize]; CHUNK_SIZE as usize],
-    fluid: Fluid,
-    x: u32,
-    y: u32,
-) {
-    let buoyancy = fluid.get_buoyancy();
-
-    // Skip if buoyancy is 0 (no movement)
-    if buoyancy == 0 {
-        // Keep the fluid in place
-        new_cells[x as usize][y as usize] = Some(Particle::Fluid(fluid));
-        return;
-    }
-
-    // Determine the vertical direction and check boundaries
-    let new_y_i32 = y as i32 + buoyancy;
-    // Use max to ensure new_y is at least 0
-    let new_y = (new_y_i32.max(0)) as u32;
-
-    // Check if we're at a vertical boundary
-    if !is_position_valid(x as i32, new_y as i32) {
-        // Keep the fluid in place at the boundary
-        new_cells[x as usize][y as usize] = Some(Particle::Fluid(fluid));
-        return;
-    }
-
-    let new_y = new_y as usize;
-
-    // Try to move in one of three directions based on priority
-    // Note: We check the original cells for obstacles, but write to new_cells
-    if is_cell_available(original_cells, new_cells, x as usize, new_y) {
-        // Move vertically
-        new_cells[x as usize][new_y] = Some(Particle::Fluid(fluid));
-    } else if x > 0
-        && is_position_valid_and_available(original_cells, new_cells, (x - 1) as i32, new_y as i32)
-    {
-        // Move diagonally to the left
-        new_cells[(x - 1) as usize][new_y] = Some(Particle::Fluid(fluid));
-    } else if is_position_valid_and_available(
-        original_cells,
-        new_cells,
-        (x + 1) as i32,
-        new_y as i32,
-    ) {
-        // Move diagonally to the right
-        new_cells[(x + 1) as usize][new_y] = Some(Particle::Fluid(fluid));
-    } else {
-        // If fluid can't move in any of these directions, it stays in place
-        new_cells[x as usize][y as usize] = Some(Particle::Fluid(fluid));
     }
 }
